+++ src/main/java/pl/cyfronet/fid/cmdbproxy/pdp/EntityStructure.java
@@ -1,5 +1,6 @@
 package pl.cyfronet.fid.cmdbproxy.pdp;
 
-interface EntityStructure {
+public interface EntityStructure {
     Entity getEntity(String entityName);
+    boolean isRoot(String entityName);
 }